Add daemon mode
In this mode all exceptions are ignored and mqtt2influxdb is reexecuted
after some amount of time to try reconnect. The idea is that when
everything is configured correctly but either mqtt server or influxdb is
down then we want to wait and not exit because otherwise execution would
be restart but only by limited number of times by daemon manager (such
as systemd or procd). The result if -d is not used is that when outage
takes more than few minutes the daemon manager notes this execution as
broken and stops restarting service. Not exiting solves that problem.
This is not perfect of course. It would be better to catch only relevant
exceptions but unless there is some bug it should be safer to just catch
every exception.